solve 14 idk
